third_party/WebKit/LayoutTests/inspector-enabled/sources/debugger/script-formatter-breakpoints-1-expected.txt

Tests the script formatting is working fine with breakpoints.


Running: testSetup

Running: testBreakpointsInOriginalAndFormattedSource
Script execution paused.
Breakpoint sidebar pane while paused in raw
script-formatter-breakpoints-1.html:12  var d = 4;
Script execution resumed.
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed
script-formatter-breakpoints-1.html:formatted:11                var d = 4;
Breakpoint sidebar pane while paused in raw
script-formatter-breakpoints-1.html:12  var d = 4;
Script execution resumed.
